@@ -1,25 +1,14 @@
-import {
-  paths,
-  parseConfig,
-  isTag,
-  unmatchedPatterns,
-  uploadUrl,
-} from "./util";
-import { release, upload, GitHubReleaser } from "./github";
-import { getOctokit } from "@actions/github";
-import { setFailed, setOutput } from "@actions/core";
-import { GitHub, getOctokitOptions } from "@actions/github/lib/utils";
+import { setFailed, setOutput } from '@actions/core';
+import { getOctokit } from '@actions/github';
+import { GitHubReleaser, release, upload } from './github';
+import { isTag, parseConfig, paths, unmatchedPatterns, uploadUrl } from './util';
 
-import { env } from "process";
+import { env } from 'process';
 
 async function run() {
   try {
     const config = parseConfig(env);
-    if (
-      !config.input_tag_name &&
-      !isTag(config.github_ref) &&
-      !config.input_draft
-    ) {
+    if (!config.input_tag_name && !isTag(config.github_ref) && !config.input_draft) {
       throw new Error(`⚠️ GitHub Releases requires a tag`);
     }
     if (config.input_files) {
@@ -45,9 +34,7 @@ async function run() {
       //new oktokit(
       throttle: {
         onRateLimit: (retryAfter, options) => {
-          console.warn(
-            `Request quota exhausted for request ${options.method} ${options.url}`
-          );
+          console.warn(`Request quota exhausted for request ${options.method} ${options.url}`);
           if (options.request.retryCount === 0) {
             // only retries once
             console.log(`Retrying after ${retryAfter} seconds!`);
@@ -56,9 +43,7 @@ async function run() {
         },
         onAbuseLimit: (retryAfter, options) => {
           // does not retry, only logs a warning
-          console.warn(
-            `Abuse detected for request ${options.method} ${options.url}`
-          );
+          console.warn(`Abuse detected for request ${options.method} ${options.url}`);
         },
       },
     });
@@ -68,33 +53,38 @@ async function run() {
       const files = paths(config.input_files);
       if (files.length == 0) {
         if (config.input_fail_on_unmatched_files) {
-          throw new Error(`⚠️ ${config.input_files} not include valid file.`);
+          throw new Error(`⚠️ ${config.input_files} does not include a valid file.`);
         } else {
-          console.warn(`🤔 ${config.input_files} not include valid file.`);
+          console.warn(`🤔 ${config.input_files} does not include a valid file.`);
         }
       }
       const currentAssets = rel.assets;
-      const assets = await Promise.all(
-        files.map(async (path) => {
-          const json = await upload(
-            config,
-            gh,
-            uploadUrl(rel.upload_url),
-            path,
-            currentAssets
-          );
+
+      const uploadFile = async (path) => {
+        const json = await upload(config, gh, uploadUrl(rel.upload_url), path, currentAssets);
+        if (json) {
           delete json.uploader;
-          return json;
-        })
-      ).catch((error) => {
-        throw error;
-      });
-      setOutput("assets", assets);
+        }
+        return json;
+      };
+
+      let results: (any | null)[];
+      if (!config.input_preserve_order) {
+        results = await Promise.all(files.map(uploadFile));
+      } else {
+        results = [];
+        for (const path of files) {
+          results.push(await uploadFile(path));
+        }
+      }
+
+      const assets = results.filter(Boolean);
+      setOutput('assets', assets);
     }
     console.log(`🎉 Release ready at ${rel.html_url}`);
-    setOutput("url", rel.html_url);
-    setOutput("id", rel.id.toString());
-    setOutput("upload_url", rel.upload_url);
+    setOutput('url', rel.html_url);
+    setOutput('id', rel.id.toString());
+    setOutput('upload_url', rel.upload_url);
   } catch (error) {
     setFailed(error.message);
   }
